For #28265 - Refactor InactiveTabs interactor/controller into TabsTray
