+++ app/src/main/java/org/mozilla/fenix/tabstray/TabsTrayController.kt
@@ -29,19 +29,24 @@ import org.mozilla.fenix.HomeActivity
 import org.mozilla.fenix.R
 import org.mozilla.fenix.browser.browsingmode.BrowsingMode
 import org.mozilla.fenix.browser.browsingmode.BrowsingModeManager
+import org.mozilla.fenix.components.AppStore
+import org.mozilla.fenix.components.appstate.AppAction
 import org.mozilla.fenix.ext.DEFAULT_ACTIVE_DAYS
+import org.mozilla.fenix.ext.potentialInactiveTabs
 import org.mozilla.fenix.home.HomeFragment
 import org.mozilla.fenix.selection.SelectionHolder
+import org.mozilla.fenix.tabstray.browser.InactiveTabsController
 import org.mozilla.fenix.tabstray.browser.SelectTabUseCaseWrapper
 import org.mozilla.fenix.tabstray.ext.isActiveDownload
 import org.mozilla.fenix.tabstray.ext.isSelect
+import org.mozilla.fenix.utils.Settings
 import java.util.concurrent.TimeUnit
 import org.mozilla.fenix.GleanMetrics.Tab as GleanTab
 
 /**
  * Controller for handling any actions in the tabs tray.
  */
-interface TabsTrayController : SyncedTabsController {
+interface TabsTrayController : SyncedTabsController, InactiveTabsController {
 
     /**
      * Called to open a new tab.
@@ -170,7 +175,9 @@ interface TabsTrayController : SyncedTabsController {
  * Default implementation of [TabsTrayController].
  *
  * @property activity [HomeActivity] used to perform top-level app actions.
+ * @property appStore [AppStore] used to dispatch any [AppAction].
  * @property tabsTrayStore [TabsTrayStore] used to read/update the [TabsTrayState].
+ * @property settings [Settings] used to update any user preferences.
  * @property browserStore [BrowserStore] used to read/update the current [BrowserState].
  * @property browsingModeManager [BrowsingModeManager] used to read/update the current [BrowsingMode].
  * @property navController [NavController] used to navigate away from the tabs tray.
@@ -185,8 +192,10 @@ interface TabsTrayController : SyncedTabsController {
 @Suppress("TooManyFunctions", "LongParameterList")
 class DefaultTabsTrayController(
     private val activity: HomeActivity,
+    private val appStore: AppStore,
     private val tabsTrayStore: TabsTrayStore,
     private val browserStore: BrowserStore,
+    private val settings: Settings,
     private val browsingModeManager: BrowsingModeManager,
     private val navController: NavController,
     private val navigateToHomeAndDeleteSession: (String) -> Unit,
@@ -423,5 +432,53 @@ class DefaultTabsTrayController(
             return true
         }
         return false
+    }
+
+    override fun handleInactiveTabClicked(tab: TabSessionState) {
+        TabsTray.openInactiveTab.add()
+        handleTabSelected(tab, TrayPagerAdapter.INACTIVE_TABS_FEATURE_NAME)
+    }
+
+    override fun handleCloseInactiveTabClicked(tab: TabSessionState) {
+        TabsTray.closeInactiveTab.add()
+        handleTabDeletion(tab.id, TrayPagerAdapter.INACTIVE_TABS_FEATURE_NAME)
+    }
+
+    override fun handleInactiveTabsHeaderClicked(expanded: Boolean) {
+        appStore.dispatch(AppAction.UpdateInactiveExpanded(expanded))
+
+        when (expanded) {
+            true -> TabsTray.inactiveTabsExpanded.record(NoExtras())
+            false -> TabsTray.inactiveTabsCollapsed.record(NoExtras())
+        }
+    }
+
+    override fun handleInactiveTabsAutoCloseDialogDismiss() {
+        markDialogAsShown()
+        TabsTray.autoCloseDimissed.record(NoExtras())
+    }
+
+    override fun handleEnableInactiveTabsAutoCloseClicked() {
+        markDialogAsShown()
+        settings.closeTabsAfterOneMonth = true
+        settings.closeTabsAfterOneWeek = false
+        settings.closeTabsAfterOneDay = false
+        settings.manuallyCloseTabs = false
+        TabsTray.autoCloseTurnOnClicked.record(NoExtras())
+    }
+
+    override fun handleDeleteAllInactiveTabsClicked() {
+        TabsTray.closeAllInactiveTabs.record(NoExtras())
+        browserStore.state.potentialInactiveTabs.map { it.id }.let {
+            tabsUseCases.removeTabs(it)
+        }
+        showUndoSnackbarForTab(false)
+    }
+
+    /**
+     * Marks the inactive tabs auto close dialog as shown and to not be displayed again.
+     */
+    private fun markDialogAsShown() {
+        settings.hasInactiveTabsAutoCloseDialogBeenDismissed = true
     }
 }
