+++ app/src/main/java/org/mozilla/fenix/tabstray/TabsTrayInteractor.kt
@@ -8,11 +8,12 @@ import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.storage.sync.Tab
 import mozilla.components.browser.tabstray.TabsTray
 import org.mozilla.fenix.selection.SelectionHolder
+import org.mozilla.fenix.tabstray.browser.InactiveTabsInteractor
 
 /**
  * Interactor for responding to all user actions in the tabs tray.
  */
-interface TabsTrayInteractor : SyncedTabsInteractor, TabsTray.Delegate {
+interface TabsTrayInteractor : SyncedTabsInteractor, TabsTray.Delegate, InactiveTabsInteractor {
     /**
      * Invoked when a page in the tabs tray is selected.
      *
@@ -186,5 +187,47 @@ class DefaultTabsTrayInteractor(
 
     override fun onTabUnselected(tab: TabSessionState) {
         controller.handleTabUnselected(tab)
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onInactiveTabsHeaderClicked].
+     */
+    override fun onInactiveTabsHeaderClicked(expanded: Boolean) {
+        controller.handleInactiveTabsHeaderClicked(expanded)
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onAutoCloseDialogCloseButtonClicked].
+     */
+    override fun onAutoCloseDialogCloseButtonClicked() {
+        controller.handleInactiveTabsAutoCloseDialogDismiss()
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onEnableAutoCloseClicked].
+     */
+    override fun onEnableAutoCloseClicked() {
+        controller.handleEnableInactiveTabsAutoCloseClicked()
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onInactiveTabClicked].
+     */
+    override fun onInactiveTabClicked(tab: TabSessionState) {
+        controller.handleInactiveTabClicked(tab)
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onInactiveTabClosed].
+     */
+    override fun onInactiveTabClosed(tab: TabSessionState) {
+        controller.handleCloseInactiveTabClicked(tab)
+    }
+
+    /**
+     * See [InactiveTabsInteractor.onDeleteAllInactiveTabsClicked].
+     */
+    override fun onDeleteAllInactiveTabsClicked() {
+        controller.handleDeleteAllInactiveTabsClicked()
     }
 }
